Add player-tracking camera.
Camera tracks player exactly so the world moves when the player moves.

Adjusted player and fireball mechanics to not care so much about the
exact boundaries of the screen. The main mob spawner routine does some
math on the camera transform to spawn mobs at the perimeter still (I
think the math works but it might be wrong once/if we add zooming).